refactor: move `fdisk` utils to `utils`
