+++ src/utils/lib.rs
@@ -1,11 +1,16 @@
-//! Features utils to several commands.
+//! Global utilities for all commands.
+
+#![feature(iter_array_chunks)]
 
 use std::env::ArgsOs;
 use std::path::PathBuf;
 use std::process::exit;
 use std::{env, fmt};
 
+pub mod crc32;
 pub mod disk;
+pub mod guid;
+pub mod partition;
 pub mod prompt;
 pub mod user;
 pub mod util;
